Dashboard styling (#707)
* confirm when user deletes files

* rename files

* fix lint errors

* Add unit test for renamePage controller method

* Add unit tests for renamePage and re-add code for checking if user owns page before deleting it

+++ client/src/app/components/Dashboard/Trash/Trash.jsx
@@ -30,11 +30,15 @@ class Trash extends React.Component {
   }
 
   deletePage = (id) => {
-    this.props.deletePage(id);
+    if (confirm('Are you sure you want to permanently delete this file?')) { // eslint-disable-line no-restricted-globals
+      this.props.deletePage(id);
+    }
   }
 
   emptyTrash = () => {
-    this.props.emptyTrash();
+    if (confirm('Are you sure you want to empty the trash?')) { // eslint-disable-line no-restricted-globals
+      this.props.emptyTrash();
+    }
   }
 
   renderTrashPages=() => {
